feat, css: Create function "addWordToDictionary" to add translation to user's dicitonary with AJAX request.
search.js: Function send AJAX request to add translation to user dictionary.
If the translation was added to the dictionary, a popover menu with translations would be closed, if it wasn't
an error message will be despalyed at the top of the page.

style.css: Add css for the error message.