feat: Delete "TranslateAPI" class and create "TranslationStrategy" abstract class to implement stragy pattern for translation.
Method of the class:

query_translation(): an abstract method to get translations for
the word from the source language to the target language.

create_templated_translations(): an abstract method to create
a list of templated translations.

translate(): common method for all strategies. Implement basic steps
to translate a word.

get_translation_template(): retur dictionary with common keys for all translations.