feat: Introduce "Translator" context class for translation strategy.
The instance of this class takes a list of strategies and put it in
"_strategies" instance variable.

translate(): method invoke a translate method in all strategies.
If the strategy method retun any translations and them to "templated_data"
dictionary and return the dictioanry.

Allso the "translate()" method add translations to database.