html: Add "htmx" implementation to "dictionary.html" and create "dictionary_word_list.html".
dictionary.html: set "htmx" attributes to"#word" input element to search words in dictionary on typing and on page load.

dictionary_word_list.html: represent a html code that contains dictonary serch of word and  will be taken and inserted to
"dictionary.html" to show words that was searched or all word of a dictionary.